Created a new `proximity.py` file
Refactoring from scratch!
